requirements.txt
aiogram==3.22.0
aiohttp>=3.9.0,<3.13
aiosqlite>=0.19.0
beautifulsoup4>=4.12.3
lxml>=5.1.0
python-dotenv>=1.0.1
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

